+++ rspec-core/spec/integration/spec_file_load_errors_spec.rb
@@ -248,7 +248,7 @@ RSpec.describe 'Spec file load errors' do
         end
       end
     else
-      it 'prints a basic error when no syntax_suggest is available/loaded', :skip => RUBY_VERSION.to_f < 1.9 || RSpec::Support::Ruby.jruby? do
+      it 'prints a basic error when no syntax_suggest is available/loaded', :skip => RSpec::Support::Ruby.jruby? do
         run_command "--require ./broken_file"
         expect(last_cmd_exit_status).to eq(error_exit_code)